+++ tests/test_dataset.py
@@ -2,6 +2,7 @@ import pandas as pd
 import pytest
 import torch
 
+from kiltergpt.data.datamodules import KilterDataModule
 from kiltergpt.data.datasets import KilterGPTDataset
 from kiltergpt.data.tokenizer import Tokenizer
 
@@ -12,12 +13,18 @@ def tokenizer():
 
 
 @pytest.fixture
-def dataset(tokenizer):
+def dataset_dir(tmp_path):
     # Create a sample dataframe
     data = {"frames": ["p1234r12p1211r13p1333r13p1421r15", "p1200r14"], "angle": [10, 20], "font_grade": ["7a", "7b"]}
     df = pd.DataFrame(data)
-    df.to_csv("data/sample.csv", index=False)
-    return KilterGPTDataset("data/sample.csv", tokenizer)
+    for i in ["train", "val", "test"]:
+        df.to_csv(tmp_path / f"{i}.csv", index=False)
+    return tmp_path
+
+
+@pytest.fixture
+def dataset(dataset_dir):
+    return KilterGPTDataset(dataset_dir / "train.csv", Tokenizer(), context_len=64)
 
 
 def test_dataset_length(dataset):
@@ -48,5 +55,8 @@ def test_label_smoothing(dataset):
     # assert (nopad[torch.isin(nopad, dataset.tokenizer.hold_token_ids())][:-1] > 1).all()
 
 
-if __name__ == "__main__":
-    pytest.main()
+def test_datamodule(dataset_dir):
+    datamodule = KilterDataModule(data_dir=dataset_dir, batch_size=2, num_workers=0)
+    datamodule.setup()
+    batch = next(iter(datamodule.train_dataloader()))
+    assert len(batch) == 2