+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a `CHANGELOG.md` file to keep track of changes in the project.
 - Added changelog-enforcer github action for pull request checks.
+- Added DataModule tests.
 
 ### Changed
 
@@ -21,5 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed paths in `preprocess.py` to be OS agnostic.
 - Removed hard-coded devices count in `train.py`.
 - Fixed paths in other scripts to be OS agnostic.
+- Fixed DataModule `num_workers` not being used.
 
 ### Deprecated